requirements.txt
plotly==4.4.1
PyYAML==5.3
